MongoUpdate refactored, added more docs
