+++ commons-mongo/src/main/scala/com/avsystem/commons/mongo/typed/MongoWrite.scala
@@ -3,6 +3,12 @@ package mongo.typed
 
 import com.mongodb.client.model._
 
+/**
+  * Represents a single MongoDB write operation in a
+  * [[https://docs.mongodb.com/manual/core/bulk-write-operations/ bulk write operation]].
+  *
+  * @tparam E type of the entity
+  */
 sealed trait MongoWrite[E] {
 
   import MongoWrite._
@@ -29,13 +35,13 @@ object MongoWrite {
 
   final case class UpdateOne[E](
     filter: MongoDocumentFilter[E],
-    update: MongoUpdate[E],
+    update: MongoDocumentUpdate[E],
     setupOptions: UpdateOptions => UpdateOptions = identity
   ) extends MongoWrite[E]
 
   final case class UpdateMany[E](
     filter: MongoDocumentFilter[E],
-    update: MongoUpdate[E],
+    update: MongoDocumentUpdate[E],
     setupOptions: UpdateOptions => UpdateOptions = identity
   ) extends MongoWrite[E]
